requirements.txt
aiofiles==0.8.0
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
asyncpg==0.25.0
attrs==21.4.0
Authlib==0.15.5
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.2
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.86.0
fastapi-restful==0.4.3
greenlet==1.1.2
h11==0.12.0
httpcore>=0.15.0
httpx==0.23.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.1
Mako==1.2.2
MarkupSafe==2.1.0
numpy==1.22.3
oauthlib==3.2.1
packaging==22.0
pandas==1.4.1
pluggy==1.0.0
psutil==5.9.0
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
pyjwt
pyparsing==3.0.7
pytest==7.1.0
pytest-asyncio==0.18.2
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.9
uvicorn==0.17.6
geojson==3.0.0
motor==3.1.2
beanie==1.17.0
aiohttp==3.8.3
rocketry==2.5.1
hsmodels
fuzzywuzzy